+++ iOS/HeartLink/HeartLink/Views/MainNavView.swift
@@ -17,7 +17,7 @@ enum PageActions: Hashable {
 struct MainNavView: View {
     @State var path: [PageActions] = [.login]
     @StateObject var btmanager = BluetoothManager()
-    @State var patient: User = User(email: "", patientId: 0, physicianId: 0, widgets: [])
+    @State var patient: User = User(email: "", patientId: 0, physicianId: 0)
     @State var recordingData: RecordingData = RecordingData(id: 0, date: "0000-00-00", viewStatus: "notSubmitted" , comments: "", fileURL: "")
     
     var body: some View {