+++ CHANGELOG.md
@@ -1,5 +1,8 @@
 # CHANGELOG
 
+## 1.7.3 (2025-01-27)
+- add confirmation dialog to app before submitting or deleting a recording
+
 ## 1.7.2 (2025-01-26)
 - add submit and delete buttons for recordings
 - separate recording history list from User model